more bugfixes
