+++ conf/blacklist.txt
@@ -3,3 +3,9 @@
 01000a10041eb005
 01000a10041eb008
 01000a10041eb003
+010028D005A3D001
+010028D005A3D002
+010028D005A3D003
+010028D005A3D004
+010028D005A3D005
+010028D005A3D008
